pyproject.toml
[project]
name = "law-crawler-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.45.1",
    "streamlit-aggrid>=1.1.5.post1",
]
